Refactor useStyles into related components files
