--- src/components/Footer.js/FooterStyles.js
@@ -1,10 +1,0 @@
-import { makeStyles } from "@material-ui/core/styles";
-
-const useStyles = makeStyles((theme) => ({
-  footer: {
-    backgroundColor: theme.palette.background.paper,
-    padding: "50px 0",
-  },
-}));
-
-export default useStyles;
