--- src/components/MainNav/MainNavStyles.js
@@ -1,15 +1,0 @@
-import { makeStyles } from "@material-ui/core";
-
-const useStyles = makeStyles((theme) => ({
-  icon: {
-    marginRight: "20px",
-  },
-  button: {
-    background: 'lightblue',
-    color: 'black',
-    margin: '8px',
-    width: 'auto'    
-  }
-}));
-
-export default useStyles;
