+++ src/components/PictureCard/PictureCardsIndex.js
@@ -2,7 +2,16 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { Container, Grid } from "@material-ui/core";
 import PictureCard from "./PictureCard";
-import useStyles from './PictureCardStyles';
+import { makeStyles } from "@material-ui/core/styles";
+
+const useStyles = makeStyles(() => ({
+  cardGrid: {
+    padding: "20px 0",
+    height: "100%",
+    display: "flex",
+    flexDirection: "column",
+  }
+}));
 
 const PictureCardsIndex = ({ cards }) => {
   const classes = useStyles();
