--- src/components/Welcome/WelcomeStyles.js
@@ -1,16 +1,0 @@
-import { makeStyles } from "@material-ui/core/styles";
-
-const useStyles = makeStyles((theme) => ({
-  container: {
-    backgroundColor: theme.palette.background.paper,
-    padding: theme.spacing(8, 0, 6),
-  },
-  icon: {
-    marginRight: "20px",
-  },
-  buttons: {
-    marginTop: "40px",
-  }
-}));
-
-export default useStyles;
